Refactored everything into sub-directories.
