+++ source/scripts/SpawnWave.js
@@ -1,12 +1,12 @@
 var Pixi = require("pixi.js")
-import Reference from "./Reference.js"
-import Junkership from "./Junkership.js"
-import Utility from "./Utility.js"
-import Trashbot from "./Trashbot.js"
-import SnakeTrashbot from "./SnakeTrashbot.js"
-import TankTrashbot from "./TankTrashbot.js"
-import SniperTrashbot from "./SniperTrashbot.js"
-import TurretTrashbot from "./TurretTrashbot.js"
+import Reference from "scripts/data/Reference.js"
+import Junkership from "scripts/sprites/Junkership.js"
+import Utility from "scripts/data/Utility.js"
+import Trashbot from "scripts/sprites/Trashbot.js"
+import SnakeTrashbot from "scripts/sprites/SnakeTrashbot.js"
+import TankTrashbot from "scripts/sprites/TankTrashbot.js"
+import SniperTrashbot from "scripts/sprites/SniperTrashbot.js"
+import TurretTrashbot from "scripts/sprites/TurretTrashbot.js"
 var Move = Trashbot.MovementStrategy
 var Shoot = Trashbot.ShootStrategy
