+++ source/scripts/sprites/Particle.js
