+++ source/scripts/sprites/SniperTrashbot.js
@@ -1,7 +1,7 @@
 var Victor = require("victor")
-import Trashbot from "./Trashbot.js"
-import Reference from "./Reference.js"
-import Projectile from "./Projectile.js"
+import Trashbot from "scripts/sprites/Trashbot.js"
+import Reference from "scripts/data/Reference.js"
+import Projectile from "scripts/sprites/Projectile.js"
 
 export default class SniperTrashbot extends Trashbot {
     constructor(position, movementStyle, shootStyle) {
