+++ source/scripts/sprites/StarStreak.js
@@ -1,7 +1,7 @@
 var Pixi = require("pixi.js")
 
-import Reference from "./Reference.js"
-import Utility from "./Utility.js"
+import Reference from "scripts/data/Reference.js"
+import Utility from "scripts/data/Utility.js"
 
 export default class StarStreak extends Pixi.Sprite {
     constructor() {
